Move runtime metrics monitoring code to a separate place
Move runtime metrics monitoring code to a separate place
